pyproject.toml
[project]
name = "ynab-report"
version = "0.1.0"
description = ""
authors = [
    {name = "Ethan Fuerst", email = "34084535+ethanfuerst@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2024.12.14",
    "urllib3>=2.3.0",
    "pandas>=2.2.3",
    "modal>=1.0.0",
    "duckdb>=1.4.0",
    "sqlfluff>=3.3.0",
    "gspread>=6.1.4",
    "gspread-formatting>=1.2.0",
    "emoji>=2.14.1",
    "sqlmesh[web]>=0.186.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
